Add decimal CPU support and China-friendly Dockerfile params

+++ backend/internal/models/instance.go
@@ -12,7 +12,7 @@ type Instance struct {
 	Description              *string    `db:"description" json:"description,omitempty"`
 	Type                     string     `db:"type" json:"type"`
 	Status                   string     `db:"status" json:"status"`
-	CPUCores                 int        `db:"cpu_cores" json:"cpu_cores"`
+	CPUCores                 float64    `db:"cpu_cores" json:"cpu_cores"`
 	MemoryGB                 int        `db:"memory_gb" json:"memory_gb"`
 	DiskGB                   int        `db:"disk_gb" json:"disk_gb"`
 	GPUEnabled               bool       `db:"gpu_enabled" json:"gpu_enabled"`
